+++ src/main/java/net/silentchaos512/gems/block/flowerpot/PhantomLightBlock.java
@@ -23,7 +23,8 @@ public class PhantomLightBlock extends Block {
     public PhantomLightBlock() {
         super(Properties.create(Material.MISCELLANEOUS)
                 .hardnessAndResistance(0.5f, 6000000)
-                .lightValue(15));
+                .lightValue(15)
+                .noDrops());
     }
 
     @Override